% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_eg}
\alias{a_eg}
\title{Example Of Interaction Coefficients Matrix}
\format{
A numeric matrix with 4 rows and 4 columns containing interaction coefficients.
}
\source{
Data generated in-house to serve as an example
}
\usage{
a_eg
}
\description{
A square numeric matrix representing interaction coefficients between
species. \code{a_ij} is the per-capita interaction strength of species \code{j}
on species \code{i}. It expresses the change in carrying capacity of species \code{i}
by a single individual of species \code{j}.
This data is compatible with \code{\link{n1_map_eg.tif}} and \code{\link{K_map_eg.tif}} maps.
}
\keyword{datasets}
