% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_eg}
\alias{community_eg}
\title{Example Community Data}
\format{
An object of class \code{sim_com_data} from the \code{mrangr} package generated
using \code{\link{initialise_com}}.
}
\source{
Data generated in-house to serve as an example
}
\description{
A pre-initialized \code{sim_com_data} object used to demonstrate community
structure and simulation input. It contains 4 species with spatially
correlated carrying capacity and initial abundance maps.

This object can be accessed via the \code{\link{get_community}} function.
}
\seealso{
\code{\link{get_community}}, \code{\link{initialise_com}}
}
