% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagonal}
\alias{diagonal}
\title{Compute Maximum Possible Distance for a Raster Object}
\usage{
diagonal(x)
}
\arguments{
\item{x}{A raster object.}
}
\value{
The diagonal distance in meteres.
}
\description{
Calculates the diagonal length of a raster's extent, accounting for the
coordinate reference system.
}
\examples{
library(terra)

# Read data from the mrangr package
K_map <- rast(system.file("input_maps/K_map_eg.tif", package = "mrangr"))

diagonal(K_map)

}
