% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_community}
\alias{get_community}
\title{Load Example Community Object}
\usage{
get_community()
}
\value{
An object of class \code{sim_com_data} containing community structure,
simulation parameters, species-specific carrying capacity and initial
abundance maps.
}
\description{
Loads a pre-simulated example of a spatial community object, useful for
demos and testing.
}
\examples{
community <- get_community()
summary(community)

}
\seealso{
\link{community_eg}
}
