% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_simulated_com}
\alias{get_simulated_com}
\title{Load Example Simulated Community Results}
\usage{
get_simulated_com()
}
\value{
An object of class \code{sim_com_results} containing simulation output for
a community over time.
}
\description{
Loads a pre-run simulation output, based on the example community data.
Useful for examples, unit tests, or visualization.
}
\examples{
sim <- get_simulated_com()
plot(sim)

}
