% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_map_eg.tif}
\alias{K_map_eg.tif}
\title{Example Of Carrying Capacity Map}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 4 layers, each
with 15 rows and 15 columns. Contains numeric values representing carrying
capacity and NA's indicating unsuitable areas.
}
\source{
Data generated in-house to serve as an example
(using spatial autocorrelation).
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 4 layer that can be
passed to \code{\link{initialise_com}} a as simulation (\code{\link{sim_com}}) starting point.

This map is compatible with \code{\link{n1_map_eg.tif}}.
}
\examples{
terra::rast(system.file("input_maps/K_map_eg.tif", package = "mrangr"))

}
