% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sim_com_data.R
\name{print.sim_com_data}
\alias{print.sim_com_data}
\title{Print \code{sim_com_data} Object}
\usage{
\method{print}{sim_com_data}(x, ...)
}
\arguments{
\item{x}{\code{sim_com_data} object; returned by the \code{\link{initialise_com}} function}

\item{...}{further arguments passed to or from other methods;
currently none specified}
}
\value{
\code{sim_com_data} object is invisibly returned (the \code{x} param)
}
\description{
Print \code{sim_com_data} Object
}
\examples{
# Read community data from the mrangr package
community <- get_community()

# Print
print(community)

}
