% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sim_com_results.R
\name{print.sim_com_results}
\alias{print.sim_com_results}
\title{Print \code{sim_com_results} Object}
\usage{
\method{print}{sim_com_results}(x, ...)
}
\arguments{
\item{x}{\code{sim_com_results} object; returned by the \code{\link{sim_com}} function}

\item{...}{further arguments passed to or from other methods; none specified}
}
\value{
\code{sim_com_results} object is invisibly returned (the \code{x} param)
}
\description{
Print \code{sim_com_results} Object
}
\examples{
# Read simulation data from the mrangr package
simulated_com <- get_simulated_com()

# Print
print(simulated_com)

}
