% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sim_com_data.R
\name{print.summary.sim_com_data}
\alias{print.summary.sim_com_data}
\title{Print \code{summary.sim_com_data} Object}
\usage{
\method{print}{summary.sim_com_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.sim_com_data}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print \code{summary.sim_com_data} Object
}
\examples{
# Read community data from the mrangr package
community <- get_community()

# Print summary
sim_com_data_summary <- summary(community)
print(sim_com_data_summary)

}
