% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sim_com_results.R
\name{print.summary.sim_com_results}
\alias{print.summary.sim_com_results}
\title{Print \code{summary.sim_results} Object}
\usage{
\method{print}{summary.sim_com_results}(x, ...)
}
\arguments{
\item{x}{\code{summary.sim_com_results} object; returned by
\code{\link{summary.sim_com_results}} function}

\item{...}{further arguments passed to or from other methods;
currently none specified}
}
\value{
None
}
\description{
Print \code{summary.sim_results} Object
}
\examples{
# Read simulation data from the mrangr package
simulated_com <- get_simulated_com()

# Print summary
sim_com_summary <- summary(simulated_com)
print(sim_com_summary)

}
