% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{to_rast}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rangr}{\code{\link[rangr]{to_rast}}}
}}

