% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_com_eg}
\alias{simulated_com_eg}
\title{Example Simulated Community Output}
\format{
An object of class \code{sim_com_results} from the \code{mrangr} package
generated using \code{\link{sim_com}}.
}
\source{
Data generated in-house to serve as an example
}
\description{
A \code{sim_com_results} object containing results of a 20-step simulation of
a 4-species community.

The simulation was generated using the \code{\link{community_eg}} object.

This object can be accessed via the \code{\link{get_simulated_com}} function.
}
\seealso{
\code{\link{get_simulated_com}}, \code{\link{plot.sim_com_results}}, \code{\link{sim_com}}
}
