% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sim_com_data.R
\name{summary.sim_com_data}
\alias{summary.sim_com_data}
\title{Summary Of \code{sim_com_data} Object}
\usage{
\method{summary}{sim_com_data}(object, ...)
}
\arguments{
\item{object}{\code{sim_com_data} object; returned by \code{\link{initialise_com}} function}

\item{...}{further arguments passed to or from other methods; currently none used}
}
\value{
A list of class \code{summary.sim_com_data}
}
\description{
Summary Of \code{sim_com_data} Object
}
\examples{
# Read community data from the mrangr package
community <- get_community()

# Summary
summary(community)

}
