% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sim_com_results.R
\name{summary.sim_com_results}
\alias{summary.sim_com_results}
\title{Summary Of \code{sim_com_results} Object}
\usage{
\method{summary}{sim_com_results}(object, ...)
}
\arguments{
\item{object}{\code{sim_com_results} object; returned by \code{\link{sim_com}} function}

\item{...}{further arguments passed to or from other methods; none specified}
}
\value{
\code{summary.sim_com_results} object
}
\description{
Summary Of \code{sim_com_results} Object
}
\examples{
# Read simulation data from the mrangr package
simulated_com <- get_simulated_com()

# Summary
summary(simulated_com)

}
