% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangr-package.R
\docType{package}
\name{mrangr-package}
\alias{mrangr}
\alias{mrangr-package}
\title{mrangr: Mechanistic Metacommunity Simulator}
\description{
A forward simulator for generating synthetic metacommunity data. As an in silico experimental platform, it enables researchers to simulate community shifts, test theoretical frameworks, and benchmark analytical algorithms prior to empirical application. Key capabilities include mechanistic simulations driven by demography, dispersal, and interactions, GIS interoperability via the 'terra' package for dynamic environments, and a virtual ecologist module that simulates imperfect detection and survey errors to mimic real-world biodiversity data.
}
\seealso{
Useful links:
\itemize{
  \item XXXX
  \item Report bugs at \url{XXXX}
}

}
\author{
\strong{Maintainer}: Author1name Author1surname \email{Author1@anonymisedemail.com} [copyright holder]

Authors:
\itemize{
  \item Author1name Author1surname \email{Author1@anonymisedemail.com} [copyright holder]
  \item Author2name Author2surname \email{Author2@anonymisedemail2.com} [copyright holder]
}

}
\keyword{internal}
