% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangr-package.R
\docType{package}
\name{mrangr-package}
\alias{mrangr}
\alias{mrangr-package}
\title{mrangr: Mechanistic Metacommunity Simulator}
\description{
A forward simulator for generating synthetic metacommunity data. As an in silico experimental platform, it enables researchers to simulate community shifts, test theoretical frameworks, and benchmark analytical algorithms prior to empirical application. Key capabilities include mechanistic simulations driven by demography, dispersal, and interactions, GIS interoperability via the 'terra' package for dynamic environments, and a virtual ecologist module that simulates imperfect detection and survey errors to mimic real-world biodiversity data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://popecol.github.io/mrangr/}
  \item \url{https://github.com/popecol/mrangr}
  \item \url{https://cran.r-project.org/web/packages/mrangr/index.html}
  \item Report bugs at \url{https://github.com/popecol/mrangr/issues}
}

}
\author{
\strong{Maintainer}: Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl} [copyright holder]

Authors:
\itemize{
  \item Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl} [copyright holder]
  \item Lechosław Kuczyński \email{lechu@amu.edu.pl} [copyright holder]
}

}
\keyword{internal}
