% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangr-package.R
\docType{package}
\name{mrangr-package}
\alias{mrangr}
\alias{mrangr-package}
\title{mrangr: Mechanistic Metacommunity Simulator}
\description{
Flexible, mechanistic, and spatially explicit simulator of metacommunities. It extends our previous package - 'rangr' (see \url{https://github.com/ropensci/rangr}), which implemented a mechanistic virtual species simulator integrating population dynamics and dispersal. The 'mrangr' package adds the ability to simulate multiple species interacting through an asymmetric matrix of pairwise relationships, allowing users to model all types of biotic interactions — competitive, facilitative, or neutral — within spatially explicit virtual environments. This work was supported by the National Science Centre, Poland, grant no. 2018/29/B/NZ8/00066 and the Poznań Supercomputing and Networking Centre (grant no. pl0090-01).
}
\author{
\strong{Maintainer}: Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl} [copyright holder]

Authors:
\itemize{
  \item Lechosław Kuczyński \email{lechu@amu.edu.pl} [copyright holder]
}

}
\keyword{internal}
