% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrangr-package.R
\docType{package}
\name{mrangr-package}
\alias{mrangr}
\alias{mrangr-package}
\title{mrangr: Mechanistic Metacommunity Simulator}
\description{
Flexible, mechanistic, and spatially explicit simulator of metacommunities. It extends the 'rangr' package, which implemented a mechanistic virtual species simulator integrating population dynamics and dispersal. The 'mrangr' package adds the ability to simulate multiple species interacting through an asymmetric matrix of pairwise relationships, allowing users to model all types of biotic interactions — competitive, facilitative, or neutral — within spatially explicit virtual environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{XXXX}
  \item Report bugs at \url{XXXX}
}

}
\author{
\strong{Maintainer}: Author1name Author1surname \email{Author1@anonimisedemail.com} [copyright holder]

Authors:
\itemize{
  \item Author2name Author2surname \email{Author2@anonimisedemail2.com} [copyright holder]
}

}
\keyword{internal}
